+++ src/main/util.ts
@@ -334,7 +334,7 @@ const openSystemExplorer = (filePath: string) => {
 /**
  * Put a save marker on a video, protecting it from the file monitor.
  */
-const toggleVideoProtectedDisk = async (videoPath: string) => {
+const protectVideoDisk = async (protect: boolean, videoPath: string) => {
   let metadata;
 
   try {
@@ -348,17 +348,13 @@ const toggleVideoProtectedDisk = async (videoPath: string) => {
     return;
   }
 
-  if (metadata.protected === undefined) {
-    console.info(`[Util] User protected ${videoPath}`);
-    metadata.protected = true;
+  if (protect) {
+    console.info(`[Util] User set protected ${videoPath}`);
   } else {
-    metadata.protected = !metadata.protected;
-
-    console.info(
-      `[Util] User toggled protection on ${videoPath}, now ${metadata.protected}`,
-    );
+    console.info(`[Util] User unprotected ${videoPath}`);
   }
 
+  metadata.protected = protect;
   await writeMetadataFile(videoPath, metadata);
 };
 
@@ -925,7 +921,7 @@ export {
   writeMetadataFile,
   deleteVideoDisk,
   openSystemExplorer,
-  toggleVideoProtectedDisk,
+  protectVideoDisk,
   fixPathWhenPackaged,
   getSortedVideos,
   getAvailableDisplays,